PSUSensor: Fix error for decimal part of scale factor
Customized scale factor declaration from schema in entity-manager is
declared as type "number." If a decimal number is used in PSUSensor,
it loses the decimal part when transforming the value from variant to
unsigned int, resulting in incorrect calculations.

Resolve this by using a variant to double and changing the scale factor
to double.

Tested:
- Added extra debug logs and verified that the customized scale values
  are loaded from the Entity Manager (EM) configuration on the Catalina
  system. The configured values are loaded and applied as expected.
```
root@bmc:~# journalctl -u xyz.openbmc_project.psusensor.service | grep customizedScale
Apr 20 10:55:07 bmc psusensor[22359]: Sensor name: HDDBOARD_SSD2_P12V_CURR_A, customizedScale: 200.000000
Apr 20 10:55:07 bmc psusensor[22359]: Sensor name: HDDBOARD_SSD2_P12V_PWR_W, customizedScale: 200000.000000
Apr 20 10:55:07 bmc psusensor[22359]: Sensor name: PDB_P48V_HSC2_CURR_A, customizedScale: 1053.907362
Apr 20 10:55:07 bmc psusensor[22359]: Sensor name: PDB_P48V_HSC2_PWR_W, customizedScale: 1054129.552522
Apr 20 10:55:07 bmc psusensor[22359]: Sensor name: PDB_P12V_NIC1_CURR_A, customizedScale: 54.083288
Apr 20 10:55:07 bmc psusensor[22359]: Sensor name: PDB_P12V_NIC1_PWR_W, customizedScale: 54077.438892
Apr 20 10:55:07 bmc psusensor[22359]: Sensor name: PDB_CABLE_TSENSE_POSITIVE_CURR_A, customizedScale: 50000.000000
Apr 20 10:55:07 bmc psusensor[22359]: Sensor name: HDDBOARD_SSD0_P12V_CURR_A, customizedScale: 200.000000
Apr 20 10:55:07 bmc psusensor[22359]: Sensor name: HDDBOARD_SSD0_P12V_PWR_W, customizedScale: 200000.000000
Apr 20 10:55:07 bmc psusensor[22359]: Sensor name: PDB_P48V_HSC1_CURR_A, customizedScale: 1061.796560
Apr 20 10:55:07 bmc psusensor[22359]: Sensor name: PDB_P48V_HSC1_PWR_W, customizedScale: 1062699.256111
Apr 20 10:55:07 bmc psusensor[22359]: Sensor name: PDB_P12V_SCM_CURR_A, customizedScale: 51.030823
Apr 20 10:55:07 bmc psusensor[22359]: Sensor name: PDB_P12V_SCM_PWR_W, customizedScale: 50875.050875
Apr 20 10:55:08 bmc psusensor[22359]: Sensor name: PDB_P12V_NIC0_CURR_A, customizedScale: 57.957575
Apr 20 10:55:08 bmc psusensor[22359]: Sensor name: PDB_P12V_NIC0_PWR_W, customizedScale: 58031.569174
Apr 20 10:55:08 bmc psusensor[22359]: Sensor name: PDB_CABLE_TSENSE_NEGATIVE_CURR_A, customizedScale: 50000.000000
```

- Verified sensor readings with customized scale applied, and the
  readings met expectations.

+++ src/psu/PSUSensor.hpp
@@ -28,7 +28,7 @@ class PSUSensor : public Sensor, public std::enable_shared_from_this<PSUSensor>
               std::vector<thresholds::Threshold>&& thresholds,
               const std::string& sensorConfiguration,
               const PowerState& powerState, std::string_view sensorUnits,
-              unsigned int factor, double max, double min, double offset,
+              double factor, double max, double min, double offset,
               const std::string& label, size_t tSize, double pollRate,
               const std::shared_ptr<I2CDevice>& i2cDevice);
     ~PSUSensor() override;
@@ -57,7 +57,7 @@ class PSUSensor : public Sensor, public std::enable_shared_from_this<PSUSensor>
     boost::asio::random_access_file inputDev;
     boost::asio::steady_timer waitTimer;
     std::string path;
-    unsigned int sensorFactor;
+    double sensorFactor;
     double sensorOffset;
     thresholds::ThresholdTimer thresholdTimer;
     void restartRead();
@@ -82,6 +82,6 @@ struct PSUProperty
     std::string_view labelTypeName;
     double maxReading;
     double minReading;
-    unsigned int sensorScaleFactor;
+    double sensorScaleFactor;
     double sensorOffset;
 };
